fix 'lint' npm script and add as CI step
The 'lint' npm script as originally written wasn't working because a path was never specified.  (a file or folder path is required).

Additionally, it was not linting typescript files because .ts and .tsx are not checked by default.  This enables those extensions as well.

Finally, I added a lint step to the GA build