move z-index attribute on component to be styled attribute
React components have a well-defined set of properties, and setting arbitrary properties on a component can generate a console error e.g.

"React does not recognize the X property on an element."
See the docs: https://reactjs.org/warnings/unknown-prop.html

In this case, the Tag component provided by Blueprint.JS "may" have had a zIndex property in the past, however the current blueprint.js docs show this property no longer exists.
See: https://blueprintjs.com/docs/#core/components/tag

This created a situation where the Blueprint @types still contains the zIndex prop but has not been updated to the latest version of blueprint, or so it seems.  If the two were matching, this probably would have been caught as a TS error.

This change fixes the console error, as well as the z-index styling on the component itself, allowing the tag to be placed above the other components.